Add files via upload
This Flipkart dashboard is all about real time data analysis of sales and weakness of products and brands, customers Behaviors, customers review towards the products of only April month. the data has been collected from the official website of Flipkart and stored in the SQL server for integration with power bi import mode.